CVChannel.jl

This Julia package and numerical analysis support the findings in The Communication Value of a Quantum Channel.

CVChannel.CVChannel — Module

A numerics library for evaluating the communication value of a quantum channel.

The communication value (CV) quantifies the performance of single-copy classical communication.

Features:

Convex optimization methods for certifying the communication value of a quantum channel.

Tools for certifying the non-multiplicativity of the communication value for quantum channels.

Contents:

A formal introduction to the communication value is found in the CV Background section.

Documentation for our optimization methods is found in the CV Optimizations section.

Documentation for our multiplicativity analysis tools are found in the CV Multiplicativity section.

Supporting methods are found in the Utilities section.

Quick Start

Install Julia: https://julialang.org/downloads/

Add the CVChannel.jl package from the Julia command prompt:

julia> using Pkg; Pkg.add("CVChannel")

To import the CVChannel.jl, run using CVChannel in the desired Julia file or workspace.

Numerical Analysis

The numerical analysis in this work investigates (non-)multiplicativity of the communication value over a wide range of quantum channels.

Scripts

This project uses scripts to verify and investigate the (non-)multiplicativity of different quantum channels. Scripts are found in the ./script directory and are categorized into two directories as:

./script/verify - assert a numerical fact or result.

./script/investigate - collects data for analysis.

Instructions for running scripts can be found in the README.md

Notebooks

Our analysis uses Jupyter notebooks for figures and demonstrations. Jupyter notebooks are found in the ./notebook directory. Notebooks are most conveniently viewed on github, however, instructions for running notebooks are found in README.md.

Citing

To cite this software please see CITATION.bib or Zenodo.

Contributing

If you are interested in contributing to this software, please reach out to the authors. Development instructions are found in the README.md.

Licensing

CVChannel.jl is released under the MIT License.

Acknowledgments

Development of CVChannel.jl was made possible by the advisory of Eric Chitambar and support from the Electrical and Computer Engineering and Physics departments at the University of Illinois at Urbana-Champaign.

Add funding